microservices/auth-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose
pyjwt
sqlalchemy
databases
asyncpg
passlib[bcrypt]
psycopg2-binary
pydantic[email]
bcrypt
redis